+++ lib/features/ai/service/operations/ai_stepfun_audio_policy.dart
@@ -96,6 +96,6 @@ final class AiStepFunAudioPolicy {
   static double? _boundedDouble(Object? raw, double min, double max) {
     final value = optionalDoubleFromValue(raw);
     if (value == null) return null;
-    return value.clamp(min, max).toDouble();
+    return value.clamp(min, max);
   }
 }
